refactor(helpers): split VSTPluginInstance god file into cohesive TUs
Resolves audit #48 F005. Relocate VSTPluginInstance member-function definitions out of the ~1150-line helpers/VSTPluginInstance.cpp into partial-class translation units (VST2/VST3/Editor/State), keeping the class header unchanged. The two cross-TU nested helpers (VST3MemoryStream, VST3HostContext) move into a new internal header VSTPluginInstanceInternal.h (ODR-safe, non-API). Main .cpp drops from ~1167 to ~430 lines. Pure refactor: COM refcounting, SEH guards, memory management, and the F036 init logging are preserved verbatim; only definitions relocated. Gated: Common.lib + EqualizerAPO.dll build with 0 errors.
