+++ helpers/VSTPluginInstance.cpp
@@ -1,4 +1,4 @@
-﻿/*
+/*
     This file is part of Equalizer APO, a system-wide equalizer.
     Copyright (C) 2017  Jonas Thedering
 
@@ -17,6 +17,19 @@
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
 
+// VSTPluginInstance is split across several translation units to keep each one
+// focused (audit #48 finding F005):
+//   - VSTPluginInstance.cpp        : construction/teardown, the VST2/VST3 dispatch
+//                                    accessors, prepareForProcessing, the audio
+//                                    process group, and the host callback plumbing.
+//   - VSTPluginInstance.VST2.cpp   : VST2 hosting (the AEffect host callback and
+//                                    initializeVST2 with its __try/__except guard).
+//   - VSTPluginInstance.VST3.cpp   : VST3 hosting (initialize/release, bus setup).
+//   - VSTPluginInstance.Editor.cpp : plugin editor window management.
+//   - VSTPluginInstance.State.cpp  : chunk/parameter state read/write.
+//   - VSTPluginInstanceInternal.h  : the nested helper classes VST3MemoryStream
+//                                    and VST3HostContext, shared across those units.
+
 #include "stdafx.h"
 #include <wincrypt.h>
 #include <inttypes.h>
@@ -30,96 +43,6 @@
 using namespace std;
 using namespace Steinberg;
 using namespace Steinberg::Vst;
-
-#define equalizerApoVSTID VST_FOURCC('E', 'A', 'P', 'O');
-
-vst_time_info vstTime{ 0,0,0,0,0,0,0,0,0,0,{0}, 0xFFFF };
-
-class VSTPluginInstance::VST3MemoryStream : public IBStream
-{
-public:
-	VST3MemoryStream() {}
-	VST3MemoryStream(const vector<char>& data) : data(data) {}
-
-	tresult PLUGIN_API queryInterface(const TUID iid, void** obj) override
-	{
-		QUERY_INTERFACE(iid, obj, FUnknown::iid, IBStream)
-		QUERY_INTERFACE(iid, obj, IBStream::iid, IBStream)
-		*obj = NULL;
-		return kNoInterface;
-	}
-
-	uint32 PLUGIN_API addRef() override { return InterlockedIncrement(&refCount); }
-	uint32 PLUGIN_API release() override
-	{
-		uint32 result = InterlockedDecrement(&refCount);
-		if (result == 0)
-			delete this;
-		return result;
-	}
-
-	tresult PLUGIN_API read(void* buffer, int32 numBytes, int32* numBytesRead = nullptr) override
-	{
-		int32 available = (int32)min<size_t>(numBytes, data.size() - min<size_t>(position, data.size()));
-		if (available > 0)
-			memcpy(buffer, data.data() + position, available);
-		position += available;
-		if (numBytesRead != NULL)
-			*numBytesRead = available;
-		return kResultOk;
-	}
-
-	tresult PLUGIN_API write(void* buffer, int32 numBytes, int32* numBytesWritten = nullptr) override
-	{
-		if (numBytes <= 0)
-		{
-			if (numBytesWritten != NULL)
-				*numBytesWritten = 0;
-			return kResultOk;
-		}
-		if (position + numBytes > data.size())
-			data.resize(position + numBytes);
-		memcpy(data.data() + position, buffer, numBytes);
-		position += numBytes;
-		if (numBytesWritten != NULL)
-			*numBytesWritten = numBytes;
-		return kResultOk;
-	}
-
-	tresult PLUGIN_API seek(int64 pos, int32 mode, int64* result = nullptr) override
-	{
-		int64 newPosition = 0;
-		if (mode == kIBSeekSet)
-			newPosition = pos;
-		else if (mode == kIBSeekCur)
-			newPosition = (int64)position + pos;
-		else if (mode == kIBSeekEnd)
-			newPosition = (int64)data.size() + pos;
-		if (newPosition < 0)
-			newPosition = 0;
-		position = (size_t)newPosition;
-		if (position > data.size())
-			data.resize(position);
-		if (result != NULL)
-			*result = (int64)position;
-		return kResultOk;
-	}
-
-	tresult PLUGIN_API tell(int64* pos) override
-	{
-		if (pos == NULL)
-			return kInvalidArgument;
-		*pos = (int64)position;
-		return kResultOk;
-	}
-
-	const vector<char>& getData() const { return data; }
-
-private:
-	volatile LONG refCount = 1;
-	vector<char> data;
-	size_t position = 0;
-};
 
 class EmptyVST3ParameterChanges : public IParameterChanges
 {
@@ -164,188 +87,6 @@ public:
 static EmptyVST3ParameterChanges emptyVST3ParameterChanges;
 static EmptyVST3EventList emptyVST3EventList;
 
-class VSTPluginInstance::VST3HostContext : public IHostApplication, public IComponentHandler, public IPlugFrame
-{
-public:
-	VST3HostContext(VSTPluginInstance* instance) : instance(instance) {}
-
-	tresult PLUGIN_API queryInterface(const TUID iid, void** obj) override
-	{
-		QUERY_INTERFACE(iid, obj, FUnknown::iid, IHostApplication)
-		QUERY_INTERFACE(iid, obj, IHostApplication::iid, IHostApplication)
-		QUERY_INTERFACE(iid, obj, IComponentHandler::iid, IComponentHandler)
-		QUERY_INTERFACE(iid, obj, IPlugFrame::iid, IPlugFrame)
-		*obj = NULL;
-		return kNoInterface;
-	}
-
-	uint32 PLUGIN_API addRef() override { return InterlockedIncrement(&refCount); }
-	uint32 PLUGIN_API release() override
-	{
-		uint32 result = InterlockedDecrement(&refCount);
-		if (result == 0)
-			delete this;
-		return result;
-	}
-
-	tresult PLUGIN_API getName(String128 name) override
-	{
-		wcsncpy_s((wchar_t*)name, 128, L"Equalizer APO", _TRUNCATE);
-		return kResultOk;
-	}
-
-	tresult PLUGIN_API createInstance(TUID, TUID, void**) override
-	{
-		return kNoInterface;
-	}
-
-	tresult PLUGIN_API beginEdit(ParamID) override { return kResultOk; }
-	tresult PLUGIN_API performEdit(ParamID, ParamValue) override
-	{
-		if (instance != NULL)
-			instance->onAutomate();
-		return kResultOk;
-	}
-	tresult PLUGIN_API endEdit(ParamID) override { return kResultOk; }
-	tresult PLUGIN_API restartComponent(int32) override { return kResultOk; }
-
-	tresult PLUGIN_API resizeView(IPlugView* view, ViewRect* newSize) override
-	{
-		if (view != NULL && newSize != NULL)
-		{
-			view->onSize(newSize);
-			if (instance != NULL)
-				instance->onSizeWindow(newSize->getWidth(), newSize->getHeight());
-		}
-		return kResultOk;
-	}
-
-private:
-	volatile LONG refCount = 1;
-	VSTPluginInstance* instance;
-};
-
-static intptr_t callback(struct vst_effect_t* effect, int32_t opcode, int32_t index, int64_t value, const char* ptr, float opt)
-{
-	VSTPluginInstance* instance = effect != NULL ? (VSTPluginInstance*)effect->host_internal : NULL;
-#ifdef _DEBUG
-	printf("vst: %p opcode: %d index: %d value: %" PRIdPTR " ptr: %p opt: %f host_internal: %p\n",
-		effect, opcode, index, value, ptr, opt, effect != NULL ? effect->host_internal : NULL);
-	fflush(stdout);
-#endif
-
-	switch (opcode)
-	{
-	case VST_HOST_OPCODE_VST_VERSION:
-		return VST_VERSION_2_4_0_0;
-
-	case VST_HOST_OPCODE_CURRENT_EFFECT_ID:
-		return equalizerApoVSTID;
-
-	case VST_EFFECT_OPCODE_PRODUCT_NAME:
-		strcpy_s((char*) ptr, 64, "Equalizer APO");
-		return 1;
-
-	case VST_EFFECT_OPCODE_VENDOR_VERSION:
-		return (intptr_t) (MAJOR << 24 | MINOR << 16 | REVISION << 8 | 0);
-
-	case VST_HOST_OPCODE_PIN_CONNECTED:
-		if (instance != NULL)
-			return index < instance->getUsedChannelCount() ? 0 : 1;
-		else
-			return 0;
-
-	case VST_HOST_OPCODE_IO_NEED_IDLE:
-		return effect != NULL ? effect->control(effect, VST_HOST_OPCODE_KEEPALIVE_OR_IDLE, 0, 0, NULL, 0.0f) : 0;
-
-	case VST_HOST_OPCODE_EDITOR_UPDATE:
-		return effect != NULL ? effect->control(effect, VST_EFFECT_OPCODE_EDITOR_KEEP_ALIVE, 0, 0, NULL, 0.0f) : 0;
-
-	case VST_HOST_OPCODE_GET_TIME:
-		if (instance != NULL) {
-			vstTime.sampleRate = instance->getSampleRate();
-			return (intptr_t)&vstTime;
-		}
-		return 0;
-
-	case VST_HOST_OPCODE_GET_SAMPLE_RATE:
-		if (instance != NULL)
-			return (intptr_t)instance->getSampleRate();
-		return 0;
-
-	case VST_HOST_OPCODE_GET_ACTIVE_THREAD:
-		if (instance != NULL)
-			return instance->getProcessLevel();
-		return 0;
-
-	case VST_HOST_OPCODE_LANGUAGE:
-		if (instance != NULL)
-			return instance->getLanguage();
-		return 0;
-
-	case VST_HOST_OPCODE_GET_REPLACE_OR_ACCUMULATE:
-		return 1;
-
-	case VST_HOST_OPCODE_EDITOR_RESIZE:
-		if (instance != NULL)
-		{
-			instance->onSizeWindow((int)index, (int)value);
-			return 0;
-		}
-		return 1;
-
-	case VST_HOST_OPCODE_SUPPORTS:
-		{
-			char* s = (char*)ptr;
-#ifdef _DEBUG
-			printf("VST canDo: %s\n", s);
-			fflush(stdout);
-#endif
-			if (strcmp(s, "startStopProcess") == 0 ||
-				strcmp(s, "sizeWindow") == 0)
-				return 1;
-		}
-		return 0;
-
-	case VST_HOST_OPCODE_AUTOMATE:
-		if (instance != NULL)
-			instance->onAutomate();
-		return 0;
-
-	case VST_HOST_OPCODE_PARAM_START_EDIT:
-	case VST_HOST_OPCODE_PARAM_STOP_EDIT:
-	case VST_HOST_OPCODE_KEEPALIVE_OR_IDLE:
-	case VST_HOST_OPCODE_WANT_MIDI:
-		return 0;
-	}
-
-	return 0;
-}
-
-static BOOL CALLBACK showChildWindow(HWND hWnd, LPARAM)
-{
-	ShowWindow(hWnd, SW_SHOW);
-	return TRUE;
-}
-
-static const wchar_t* vst3EditorHostWindowClass = L"EqualizerAPOVST3EditorHost";
-
-static void registerVST3EditorHostWindowClass()
-{
-	static bool registered = false;
-	if (registered)
-		return;
-
-	WNDCLASSW wc;
-	memset(&wc, 0, sizeof(wc));
-	wc.lpfnWndProc = DefWindowProcW;
-	wc.hInstance = GetModuleHandleW(NULL);
-	wc.lpszClassName = vst3EditorHostWindowClass;
-	wc.hCursor = LoadCursor(NULL, IDC_ARROW);
-	RegisterClassW(&wc);
-	registered = true;
-}
-
 VSTPluginInstance::VSTPluginInstance(const std::shared_ptr<VSTPluginLibrary>& library, int processLevel)
 	: library(library), processLevel(processLevel)
 {
@@ -377,115 +118,6 @@ bool VSTPluginInstance::initialize()
 	else if (effect == NULL)
 		LogF(L"VST2 plugin %s has wrong magic number, not loading it.", library->getLibPath().c_str());
 	return result;
-}
-
-bool VSTPluginInstance::initializeVST2()
-{
-	bool result = true;
-
-	__try
-	{
-		effect = library->VSTPluginMain(callback);
-		effect->host_internal = this;
-		if (effect->magic_number == VST_MAGICNUMBER)
-		{
-			effect->control(effect, VST_EFFECT_OPCODE_INITIALIZE, 0, 0, NULL, 0.0f);
-
-			usedChannelCount = max(numInputs(), numOutputs());
-		}
-		else
-		{
-			effect = NULL;
-		}
-	}
-	__except (EXCEPTION_EXECUTE_HANDLER)
-	{
-		result = false;
-	}
-
-	return result;
-}
-
-bool VSTPluginInstance::initializeVST3()
-{
-	vst3HostContext = new VST3HostContext(this);
-
-	const PClassInfo& classInfo = library->getVST3ClassInfo();
-	FUID componentId(classInfo.cid);
-	TUID componentIid;
-	IComponent::iid.toTUID(componentIid);
-	if (library->getFactory()->createInstance(componentId, componentIid, (void**)&vst3Component) != kResultOk || vst3Component == NULL)
-	{
-		LogF(L"Could not create IComponent instance of VST3 plugin %s.", library->getLibPath().c_str());
-		return false;
-	}
-
-	vst3Component->setIoMode(kSimple);
-	if (vst3Component->initialize(static_cast<IHostApplication*>(vst3HostContext)) != kResultOk)
-	{
-		LogF(L"Could not initialize IComponent of VST3 plugin %s.", library->getLibPath().c_str());
-		return false;
-	}
-
-	TUID processorIid;
-	IAudioProcessor::iid.toTUID(processorIid);
-	if (vst3Component->queryInterface(processorIid, (void**)&vst3Processor) != kResultOk || vst3Processor == NULL)
-	{
-		LogF(L"VST3 plugin %s does not provide the IAudioProcessor interface.", library->getLibPath().c_str());
-		return false;
-	}
-
-	TUID controllerClassId;
-	memset(controllerClassId, 0, sizeof(controllerClassId));
-	if (vst3Component->getControllerClassId(controllerClassId) == kResultOk)
-	{
-		TUID controllerIid;
-		IEditController::iid.toTUID(controllerIid);
-		library->getFactory()->createInstance(controllerClassId, controllerIid, (void**)&vst3Controller);
-	}
-	if (vst3Controller == NULL)
-	{
-		TUID controllerIid;
-		IEditController::iid.toTUID(controllerIid);
-		vst3Component->queryInterface(controllerIid, (void**)&vst3Controller);
-	}
-	if (vst3Controller != NULL)
-	{
-		vst3Controller->initialize(static_cast<IHostApplication*>(vst3HostContext));
-		vst3Controller->setComponentHandler(vst3HostContext);
-
-		VST3MemoryStream* stream = new VST3MemoryStream();
-		if (vst3Component->getState(stream) == kResultOk)
-		{
-			stream->seek(0, IBStream::kIBSeekSet);
-			vst3Controller->setComponentState(stream);
-		}
-		stream->release();
-
-		TUID connectionPointIid;
-		Steinberg::Vst::IConnectionPoint::iid.toTUID(connectionPointIid);
-		if (vst3Component->queryInterface(connectionPointIid, (void**)&vst3ComponentConnection) == kResultOk
-			&& vst3Controller->queryInterface(connectionPointIid, (void**)&vst3ControllerConnection) == kResultOk)
-		{
-			vst3ComponentConnection->connect(vst3ControllerConnection);
-			vst3ControllerConnection->connect(vst3ComponentConnection);
-		}
-	}
-
-	vst3InputBusCount = max(0, vst3Component->getBusCount(kAudio, kInput));
-	vst3OutputBusCount = max(0, vst3Component->getBusCount(kAudio, kOutput));
-	configureVST3Buses(2);
-
-	vst3SupportsDouble = vst3Processor->canProcessSampleSize(kSample64) == kResultOk;
-	if (!vst3SupportsDouble && vst3Processor->canProcessSampleSize(kSample32) != kResultOk)
-	{
-		LogF(L"VST3 plugin %s supports neither 32-bit nor 64-bit sample processing.", library->getLibPath().c_str());
-		return false;
-	}
-
-	usedChannelCount = max(numInputs(), numOutputs());
-
-	return true;
 }
 
 int VSTPluginInstance::numInputs() const
@@ -627,139 +259,6 @@ void VSTPluginInstance::prepareForProcessing(float sampleRate, int blockSize)
 	effect->control(effect, VST_EFFECT_OPCODE_SET_BLOCK_SIZE, 0, blockSize, NULL, 0.0f);
 }
 
-void VSTPluginInstance::writeToEffect(const std::wstring& chunkData, const std::unordered_map<std::wstring, float>& paramMap)
-{
-	if (library->isVST3())
-	{
-		if (chunkData != L"")
-		{
-			DWORD bufSize = 0;
-			CryptStringToBinaryW(chunkData.c_str(), 0, CRYPT_STRING_BASE64, NULL, &bufSize, NULL, NULL);
-			vector<char> data(bufSize);
-			if (bufSize > 0 && CryptStringToBinaryW(chunkData.c_str(), 0, CRYPT_STRING_BASE64, (BYTE*)data.data(), &bufSize, NULL, NULL) == TRUE)
-			{
-				VST3MemoryStream* stream = new VST3MemoryStream(data);
-				if (vst3Component != NULL)
-					vst3Component->setState(stream);
-				stream->seek(0, IBStream::kIBSeekSet);
-				if (vst3Controller != NULL)
-					vst3Controller->setState(stream);
-				stream->release();
-			}
-		}
-		else if (vst3Controller != NULL)
-		{
-			for (auto it : paramMap)
-			{
-				for (int32 i = 0; i < vst3Controller->getParameterCount(); i++)
-				{
-					ParameterInfo info;
-					if (vst3Controller->getParameterInfo(i, info) == kResultOk
-						&& it.first == wstring((wchar_t*)info.title))
-					{
-						vst3Controller->setParamNormalized(info.id, it.second);
-						break;
-					}
-				}
-			}
-		}
-		return;
-	}
-
-	if (effect == NULL)
-		return;
-
-	if (effect->flags & VST_EFFECT_FLAG_CHUNKS)
-	{
-		if (chunkData != L"")
-		{
-			DWORD bufSize = 0;
-			CryptStringToBinaryW(chunkData.c_str(), 0, CRYPT_STRING_BASE64, NULL, &bufSize, NULL, NULL);
-			BYTE* buf = new BYTE[bufSize];
-			if (CryptStringToBinaryW(chunkData.c_str(), 0, CRYPT_STRING_BASE64, buf, &bufSize, NULL, NULL) == TRUE)
-				effect->control(effect, VST_EFFECT_OPCODE_SET_CHUNK_DATA, 1, bufSize, buf, 0.0f);
-			delete[] buf;
-		}
-	}
-	else
-	{
-		for (int i = 0; i < effect->num_params; i++)
-		{
-			char buf[256];
-			effect->control(effect, VST_EFFECT_OPCODE_PARAM_GET_NAME, i, 0, buf, 0.0f);
-			buf[255] = '\0'; // just to be sure
-			wstring name = StringHelper::toWString(buf, CP_UTF8);
-			auto it = paramMap.find(name);
-			if (it != paramMap.end())
-				effect->set_parameter(effect, i, it->second);
-		}
-	}
-}
-
-void VSTPluginInstance::readFromEffect(std::wstring& chunkData, std::unordered_map<std::wstring, float>& paramMap) const
-{
-	if (library->isVST3())
-	{
-		chunkData = L"";
-		paramMap.clear();
-
-		if (vst3Component != NULL)
-		{
-			VST3MemoryStream* stream = new VST3MemoryStream();
-			if (vst3Component->getState(stream) == kResultOk && !stream->getData().empty())
-			{
-				DWORD stringLength = 0;
-				CryptBinaryToStringW((BYTE*)stream->getData().data(), (DWORD)stream->getData().size(), CRYPT_STRING_BASE64 | CRYPT_STRING_NOCRLF, NULL, &stringLength);
-				wchar_t* string = new wchar_t[stringLength];
-				if (CryptBinaryToStringW((BYTE*)stream->getData().data(), (DWORD)stream->getData().size(), CRYPT_STRING_BASE64 | CRYPT_STRING_NOCRLF, string, &stringLength) == TRUE)
-					chunkData = string;
-				delete[] string;
-			}
-			stream->release();
-		}
-
-		if (chunkData == L"" && vst3Controller != NULL)
-		{
-			for (int32 i = 0; i < vst3Controller->getParameterCount(); i++)
-			{
-				ParameterInfo info;
-				if (vst3Controller->getParameterInfo(i, info) == kResultOk)
-					paramMap[wstring((wchar_t*)info.title)] = (float)vst3Controller->getParamNormalized(info.id);
-			}
-		}
-		return;
-	}
-
-	if (effect == NULL)
-		return;
-
-	chunkData = L"";
-	paramMap.clear();
-
-	if (effect->flags & VST_EFFECT_FLAG_CHUNKS)
-	{
-		BYTE* chunk = NULL;
-		int size = (int)effect->control(effect, VST_EFFECT_OPCODE_GET_CHUNK_DATA, 1, 0, &chunk, 0.0f);
-		DWORD stringLength = 0;
-		CryptBinaryToStringW(chunk, size, CRYPT_STRING_BASE64 | CRYPT_STRING_NOCRLF, NULL, &stringLength);
-		wchar_t* string = new wchar_t[stringLength];
-		if (CryptBinaryToStringW(chunk, size, CRYPT_STRING_BASE64 | CRYPT_STRING_NOCRLF, string, &stringLength) == TRUE)
-			chunkData = string;
-		delete[] string;
-	}
-	else
-	{
-		for (int i = 0; i < effect->num_params; i++)
-		{
-			char buf[256];
-			effect->control(effect, VST_EFFECT_OPCODE_PARAM_GET_NAME, i, 0, buf, 0.0f);
-			buf[255] = '\0'; // just to be sure
-			float value = effect->get_parameter(effect, i);
-			paramMap[StringHelper::toWString(buf, CP_UTF8)] = value;
-		}
-	}
-}
-
 void VSTPluginInstance::startProcessing()
 {
 	if (library->isVST3())
@@ -898,126 +397,6 @@ void VSTPluginInstance::stopProcessing()
 	effect->control(effect, VST_EFFECT_OPCODE_SUSPEND, 0, 0, NULL, 0.0f);
 }
 
-bool VSTPluginInstance::startEditing(HWND hWnd, short* width, short* height)
-{
-	if (width != NULL)
-		*width = 400;
-	if (height != NULL)
-		*height = 300;
-
-	if (library->isVST3())
-	{
-		if (vst3Controller == NULL)
-			return false;
-		stopEditing();
-		vst3View = vst3Controller->createView(ViewType::kEditor);
-		if (vst3View == NULL)
-			return false;
-		vst3View->setFrame(vst3HostContext);
-		ViewRect rect;
-		if (vst3View->getSize(&rect) == kResultOk)
-		{
-			if (width != NULL)
-				*width = (short)max<int32>(1, rect.getWidth());
-			if (height != NULL)
-				*height = (short)max<int32>(1, rect.getHeight());
-		}
-		tresult platformResult = vst3View->isPlatformTypeSupported(kPlatformTypeHWND);
-		if (platformResult != kResultOk && platformResult != kNotImplemented)
-		{
-			stopEditing();
-			return false;
-		}
-		registerVST3EditorHostWindowClass();
-		vst3EditorHostWindow = CreateWindowExW(0, vst3EditorHostWindowClass, L"",
-			WS_CHILD | WS_CLIPCHILDREN | WS_CLIPSIBLINGS | WS_VISIBLE,
-			0, 0, *width, *height, hWnd, NULL, GetModuleHandleW(NULL), NULL);
-		if (vst3EditorHostWindow == NULL)
-		{
-			stopEditing();
-			return false;
-		}
-		if (vst3View->attached(vst3EditorHostWindow, kPlatformTypeHWND) != kResultOk)
-		{
-			stopEditing();
-			return false;
-		}
-		if (vst3View->getSize(&rect) == kResultOk)
-		{
-			if (width != NULL)
-				*width = (short)max<int32>(1, rect.getWidth());
-			if (height != NULL)
-				*height = (short)max<int32>(1, rect.getHeight());
-			vst3View->onSize(&rect);
-		}
-		SetWindowPos(vst3EditorHostWindow, NULL, 0, 0, *width, *height, SWP_NOZORDER | SWP_NOACTIVATE | SWP_SHOWWINDOW);
-		EnumChildWindows(vst3EditorHostWindow, showChildWindow, 0);
-		return true;
-	}
-
-	if (effect == NULL)
-		return false;
-
-	vst_rect_t* rect;
-	effect->control(effect, VST_EFFECT_OPCODE_EDITOR_GET_RECT, 0, 0, &rect, 0.0f);
-	effect->control(effect, VST_EFFECT_OPCODE_EDITOR_OPEN, 0, 0, hWnd, 0.0f);
-	effect->control(effect, VST_EFFECT_OPCODE_EDITOR_GET_RECT, 0, 0, &rect, 0.0f);
-
-	if (width != NULL)
-		*width = rect->right - rect->left;
-	if (height != NULL)
-		*height = rect->bottom - rect->top;
-	return true;
-}
-
-void VSTPluginInstance::doIdle()
-{
-	if (library->isVST3())
-	{
-		if (vst3EditorHostWindow != NULL)
-		{
-			InvalidateRect(vst3EditorHostWindow, NULL, FALSE);
-			UpdateWindow(vst3EditorHostWindow);
-			EnumChildWindows(vst3EditorHostWindow, [](HWND hWnd, LPARAM) -> BOOL {
-				InvalidateRect(hWnd, NULL, FALSE);
-				UpdateWindow(hWnd);
-				return TRUE;
-			}, 0);
-		}
-		return;
-	}
-
-	if (effect == NULL)
-		return;
-
-	effect->control(effect, VST_EFFECT_OPCODE_EDITOR_KEEP_ALIVE, 0, 0, NULL, 0.0f);
-}
-
-void VSTPluginInstance::stopEditing()
-{
-	if (library->isVST3())
-	{
-		if (vst3View != NULL)
-		{
-			vst3View->removed();
-			vst3View->setFrame(NULL);
-			vst3View->release();
-			vst3View = NULL;
-		}
-		if (vst3EditorHostWindow != NULL)
-		{
-			DestroyWindow(vst3EditorHostWindow);
-			vst3EditorHostWindow = NULL;
-		}
-		return;
-	}
-
-	if (effect == NULL)
-		return;
-
-	effect->control(effect, VST_EFFECT_OPCODE_EDITOR_CLOSE, 0, 0, NULL, 0.0f);
-}
-
 void VSTPluginInstance::setAutomateFunc(std::function<void()> func)
 {
 	automateFunc = func;
@@ -1047,120 +426,4 @@ void VSTPluginInstance::onSizeWindow(int w, int h)
 		SetWindowPos(vst3EditorHostWindow, NULL, 0, 0, w, h, SWP_NOZORDER | SWP_NOACTIVATE | SWP_SHOWWINDOW);
 	if (sizeWindowFunc)
 		sizeWindowFunc(w, h);
-}
-
-void VSTPluginInstance::releaseVST3()
-{
-	if (!library->isVST3())
-		return;
-
-	automateFunc = nullptr;
-	sizeWindowFunc = nullptr;
-	stopEditing();
-	stopProcessing();
-
-	if (vst3Controller != NULL)
-		vst3Controller->setComponentHandler(NULL);
-
-	if (vst3ComponentConnection != NULL && vst3ControllerConnection != NULL)
-	{
-		vst3ComponentConnection->disconnect(vst3ControllerConnection);
-		vst3ControllerConnection->disconnect(vst3ComponentConnection);
-	}
-	if (vst3ComponentConnection != NULL)
-	{
-		vst3ComponentConnection->release();
-		vst3ComponentConnection = NULL;
-	}
-	if (vst3ControllerConnection != NULL)
-	{
-		vst3ControllerConnection->release();
-		vst3ControllerConnection = NULL;
-	}
-	if (vst3Controller != NULL)
-	{
-		vst3Controller->terminate();
-		vst3Controller->release();
-		vst3Controller = NULL;
-	}
-	if (vst3Processor != NULL)
-	{
-		vst3Processor->release();
-		vst3Processor = NULL;
-	}
-	if (vst3Component != NULL)
-	{
-		vst3Component->terminate();
-		vst3Component->release();
-		vst3Component = NULL;
-	}
-	if (vst3HostContext != NULL)
-	{
-		vst3HostContext->release();
-		vst3HostContext = NULL;
-	}
-}
-
-void VSTPluginInstance::configureVST3Buses(int requestedChannelCount)
-{
-	if (vst3Component == NULL || vst3Processor == NULL)
-		return;
-
-	int channelCount = max(1, requestedChannelCount);
-	SpeakerArrangement inputArrangement = speakerArrangementForChannelCount(channelCount);
-	SpeakerArrangement outputArrangement = speakerArrangementForChannelCount(channelCount);
-
-	if (inputArrangement == SpeakerArr::kEmpty || outputArrangement == SpeakerArr::kEmpty)
-	{
-		BusInfo busInfo;
-		memset(&busInfo, 0, sizeof(busInfo));
-		if (vst3InputBusCount > 0 && vst3Component->getBusInfo(kAudio, kInput, 0, busInfo) == kResultOk && busInfo.channelCount > 0)
-			inputArrangement = speakerArrangementForChannelCount(busInfo.channelCount);
-		if (vst3OutputBusCount > 0 && vst3Component->getBusInfo(kAudio, kOutput, 0, busInfo) == kResultOk && busInfo.channelCount > 0)
-			outputArrangement = speakerArrangementForChannelCount(busInfo.channelCount);
-	}
-
-	for (int i = 0; i < vst3InputBusCount; i++)
-		vst3Component->activateBus(kAudio, kInput, i, i == 0);
-	for (int i = 0; i < vst3OutputBusCount; i++)
-		vst3Component->activateBus(kAudio, kOutput, i, i == 0);
-
-	if (inputArrangement != SpeakerArr::kEmpty || outputArrangement != SpeakerArr::kEmpty)
-	{
-		vst3Processor->setBusArrangements(
-			vst3InputBusCount > 0 && inputArrangement != SpeakerArr::kEmpty ? &inputArrangement : NULL,
-			vst3InputBusCount > 0 && inputArrangement != SpeakerArr::kEmpty ? 1 : 0,
-			vst3OutputBusCount > 0 && outputArrangement != SpeakerArr::kEmpty ? &outputArrangement : NULL,
-			vst3OutputBusCount > 0 && outputArrangement != SpeakerArr::kEmpty ? 1 : 0);
-	}
-
-	BusInfo busInfo;
-	memset(&busInfo, 0, sizeof(busInfo));
-	if (vst3InputBusCount > 0 && vst3Component->getBusInfo(kAudio, kInput, 0, busInfo) == kResultOk)
-		vst3InputChannelCount = max(0, busInfo.channelCount);
-	if (vst3OutputBusCount > 0 && vst3Component->getBusInfo(kAudio, kOutput, 0, busInfo) == kResultOk)
-		vst3OutputChannelCount = max(0, busInfo.channelCount);
-}
-
-SpeakerArrangement VSTPluginInstance::speakerArrangementForChannelCount(int count) const
-{
-	switch (count)
-	{
-	case 1:
-		return SpeakerArr::kMono;
-	case 2:
-		return SpeakerArr::kStereo;
-	case 4:
-		return SpeakerArr::k40Music;
-	case 5:
-		return SpeakerArr::k50;
-	case 6:
-		return SpeakerArr::k51;
-	case 7:
-		return SpeakerArr::k61Cine;
-	case 8:
-		return SpeakerArr::k71Music;
-	default:
-		return SpeakerArr::kEmpty;
-	}
 }
